Added method to format the Json value of transactional Delete

+++ spec/transaction.spec.js
@@ -1,99 +1,103 @@
-'use strict';
-
-
 if (typeof module !== 'undefined') {
   require('./node');
-
 }
 
-describe('Test Transaction', function() {
+describe('Test Transaction', function () {
   var emf, metamodel, rootDb;
-  before(function() {
-    var personType,studentType;
+  before(function () {
+    var personType, studentType;
 
-    emf = new DB.EntityManagerFactory({host: env.TEST_SERVER, schema: [], tokenStorage: helper.rootTokenStorage});
+    emf = new DB.EntityManagerFactory({ host: env.TEST_SERVER, schema: [], tokenStorage: helper.rootTokenStorage });
     metamodel = emf.metamodel;
     metamodel.addType(personType = new DB.metamodel.EntityType('PersonTable', metamodel.entity(Object)));
     personType.addAttribute(new DB.metamodel.SingularAttribute('name', metamodel.baseType(String)));
     personType.addAttribute(new DB.metamodel.SingularAttribute('age', metamodel.baseType(Number)));
     personType.addAttribute(new DB.metamodel.SingularAttribute('zip', metamodel.baseType(Number)));
 
-
     metamodel.addType(studentType = new DB.metamodel.EntityType('Student', metamodel.entity(Object)));
     studentType.addAttribute(new DB.metamodel.SingularAttribute('name', metamodel.baseType(String)));
     studentType.addAttribute(new DB.metamodel.SingularAttribute('address', metamodel.baseType(String)));
     studentType.addAttribute(new DB.metamodel.SingularAttribute('person', personType));
-    
 
     return metamodel.save();
   });
 
-  before(function() {
+  before(function () {
     emf = new DB.EntityManagerFactory(env.TEST_SERVER);
-    return emf.createEntityManager().ready().then(function(em) {
-      return em.User.login('root', 'root').then( function() {
+    return emf.createEntityManager().ready().then(function (em) {
+      return em.User.login('root', 'root').then(function () {
         rootDb = em;
-         rootDb.PersonTable.find().equal('name', 'helloworld1').singleResult( (data) => {
-          if(data)
-             data.delete();
+        rootDb.PersonTable.find().equal('name', 'helloworld1').singleResult((data) => {
+          if (data) data.delete();
         });
-         rootDb.PersonTable.find().equal('name', 'helloworld2').singleResult( (data) => {
-          if(data)
-             data.delete();
+        rootDb.PersonTable.find().equal('name', 'helloworld2').singleResult((data) => {
+          if (data) data.delete();
         });
-        rootDb.PersonTable.find().equal('name', 'helloworld3').singleResult( (data) => {
-          if(data)
-            data.delete();
+        rootDb.PersonTable.find().equal('name', 'helloworld3').singleResult((data) => {
+          if (data) data.delete();
         });
-         rootDb.Student.find().equal('name', 'Test Student').singleResult( (data) => {
-          if(data)
-             data.delete();
+        rootDb.Student.find().equal('name', 'Test Student').singleResult((data) => {
+          if (data) data.delete();
+        });
+        rootDb.Student.find().equal('name', 'Student_X').singleResult((data) => {
+          if (data) data.delete();
+        });
+        rootDb.Student.find().equal('name', 'person1').singleResult((data) => {
+          if (data) data.delete();
         });
       });
     });
   });
 
-  describe('Transaction Begin and Commit', function() {
-    it('Insert new records', function() {
-      return rootDb.transaction.begin().then(function(txid) {
+  describe('Transaction Begin and Commit', function () {
+    it('Insert new records', function () {
+      return rootDb.transaction.begin().then(function (txid) {
         expect(txid).to.be.not.null;
-        let tt = rootDb.PersonTable();
+        const tt = rootDb.PersonTable();
         tt.id = '890';
         tt.name = 'helloworld1';
         tt.age = 32;
         tt.zip = 5600081;
         tt.save();
 
-        let tt1 = rootDb.PersonTable();
+        const tt1 = rootDb.PersonTable();
         tt1.id = '123';
         tt1.name = 'helloworld2';
         tt1.age = 50;
         tt1.zip = 5600082;
         tt1.save();
 
-        let st = rootDb.Student();
+        const st = rootDb.Student();
         st.id = '12121';
         st.name = 'Test Student';
         st.address = 'home';
         st.person = tt1;
         st.save();
+
+        const st1 = rootDb.Student();
+        st1.id = '111';
+        st1.name = 'Student_X';
+        st1.address = 'Area_X';
+        st1.person = tt;
+        st1.save();
+
         return Promise.resolve();
-      }).then(function() {
-        return rootDb.transaction.commit().then(function(response){
+      }).then(function () {
+        return rootDb.transaction.commit().then(function (response) {
           console.log(response);
           expect(response).to.be.not.null;
-          expect(Object.keys(response).length).equals(3);
+          expect(Object.keys(response).length).equals(4);
         });
-      }).catch(function(error) {
-        console.log('ERROR: ' + JSON.stringify(error));
-        expect.fail("No Error should have been thrown");
-        //rootDb.transaction.rollback(txnObj);
+      }).catch(function (error) {
+        console.log(`ERROR: ${JSON.stringify(error)}`);
+        expect.fail('No Error should have been thrown');
+        // rootDb.transaction.rollback(txnObj);
       });
     });
 
-    it('Update existing records', function() {
-      return rootDb.transaction.begin().then( async function(txid) {
-        await rootDb.PersonTable.find().equal('name', 'helloworld1').singleResult( (data) => {
+    it('Update existing records', function () {
+      return rootDb.transaction.begin().then(async function (txid) {
+        await rootDb.PersonTable.find().equal('name', 'helloworld1').singleResult((data) => {
           if (data) {
             data.name = 'helloworld1';
             data.age = 33;
@@ -119,7 +123,7 @@ describe('Test Transaction', function() {
           }
         });
 
-        let tt = rootDb.PersonTable();
+        const tt = rootDb.PersonTable();
         tt.id = '777';
         tt.name = 'helloworld3';
         tt.age = 54;
@@ -127,72 +131,101 @@ describe('Test Transaction', function() {
         tt.save();
 
         return Promise.resolve();
-
-      }).then(function() {
-        return rootDb.transaction.commit().then(function(response){
+      }).then(function () {
+        return rootDb.transaction.commit().then(function (response) {
           console.log(response);
           expect(response).to.be.not.null;
           expect(Object.keys(response).length).equals(4);
         });
-      }).catch(function(error) {
-        console.log('ERROR: ' + JSON.stringify(error));
-        expect.fail("No Error should have been thrown here "+error);
-        //rootDb.transaction.rollback(txnObj);
+      }).catch(function (error) {
+        console.log(`ERROR: ${JSON.stringify(error)}`);
+        expect.fail(`No Error should have been thrown here ${error}`);
+        // rootDb.transaction.rollback(txnObj);
       });
     });
 
-    it('rollback created records', function() {
-      return rootDb.transaction.begin().then( async function(txid) {
+    it('Create and Delete record', function () {
+      return rootDb.transaction.begin().then(async function (txid) {
         expect(txid).to.be.not.null;
-        let tt = rootDb.PersonTable();
+
+        const p1 = rootDb.PersonTable();
+        p1.id = '101';
+        p1.name = 'person1';
+        p1.age = 36;
+        p1.zip = 210103;
+        p1.save();
+
+        await rootDb.Student.find().equal('name', 'Student_X').singleResult((data) => {
+          if (data) {
+            data.delete();
+          }
+        });
+        return Promise.resolve();
+      }).then(function () {
+        return rootDb.transaction.commit().then(function (response) {
+          console.log(response);
+          expect(response).to.be.not.null;
+          expect(Object.keys(response).length).equals(1);
+        });
+      }).catch(function (error) {
+        console.log(`ERROR: ${JSON.stringify(error)}`);
+        expect.fail(`No Error should have been thrown here ${error}`);
+        // rootDb.transaction.rollback(txnObj);
+      });
+    });
+
+    it('rollback created records', function () {
+      return rootDb.transaction.begin().then(async function (txid) {
+        expect(txid).to.be.not.null;
+        const tt = rootDb.PersonTable();
         tt.id = '890';
         tt.name = 'Umarou';
         tt.age = 32;
         tt.zip = 5600081;
         tt.save();
-        return Promise.resolve();      
-      }).then(function() {
-         //rollback with transaction ID: expected to succeed
-         rootDb.transaction.rollback().then(function(response) {
-           expect(response).to.be.empty;
-         },
-         function(failedResp) {
-           expect.fail("Wasn't expected to fail: "+failedResp);
-         });
+        return Promise.resolve();
+      }).then(function () {
+        // rollback with transaction ID: expected to succeed
+        rootDb.transaction.rollback().then(function (response) {
+          expect(response).to.be.empty;
+        },
+        function (failedResp) {
+          expect.fail(`Wasn't expected to fail: ${failedResp}`);
+        });
 
-          //rollback with no transaction ID: expected to fail
-          rootDb.transaction.rollback().then(function(response) {
-              expect.fail("Wasn't expected to succeed");            
-          },
-            function(failedResp) {
-              expect(failedResp).to.be.eq("Nothing to do. Transaction does not exist");
-          });
-                     
-        //No record with name Umarou should be found in the db since created record was rolled back        
-        return rootDb.transaction.begin().then( async function(txid) {
+        // rollback with no transaction ID: expected to fail
+        rootDb.transaction.rollback().then(function (response) {
+          expect.fail("Wasn't expected to succeed");
+        },
+        function (failedResp) {
+          expect(failedResp).to.be.eq('Nothing to do. Transaction does not exist');
+        });
+
+        // No record with name Umarou should be found in the db since created record was rolled back
+        return rootDb.transaction.begin().then(async function (txid) {
           expect(txid).to.be.not.null;
-          await rootDb.PersonTable.find().equal('name', 'Umarou').singleResult( (data) => {
+          await rootDb.PersonTable.find().equal('name', 'Umarou').singleResult((data) => {
             if (data) {
-              expect.fail("Unexpected Data");
+              expect.fail('Unexpected Data');
             }
-          });      
-        })
-      })
+          });
+        });
+      });
     });
 
-    it('No duplicate transaction check', function() {
-      return rootDb.transaction.begin().then(function(txid) {
+    it('No duplicate transaction check', function () {
+      return rootDb.transaction.begin().then(function (txid) {
         expect(txid).to.be.not.null;
-        rootDb.transaction.begin().then(function() {
-          expect.fail("Transaction begin should have failed");
-        }).catch( function(error) {
+        rootDb.transaction.begin().then(function () {
+          expect.fail('Transaction begin should have failed');
+        }).catch(function (error) {
           expect(error).to.throw(Error('Transaction already exist.. Please commit existing transaction first'));
-          return rootDb.transaction.commit().then(function(response){
+          return rootDb.transaction.commit().then(function (response) {
             console.log(response);
             expect(response).to.be.not.null;
           });
         });
-      })
+      });
     });
   });
 });
